refactor: changed test organisation structure
