--- booking/src/test/java/com/komsije/booking/e2eTests/SearchAndFilterAccommodations/SearchAndFilterTest.java
@@ -1,5 +1,0 @@
-package com.komsije.booking.e2eTests.SearchAndFilterAccommodations;
-
-public class SearchAndFilterTest {
-
-}
