+++ booking/src/test/java/com/komsije/booking/e2eTests/pages/GuestProfilePage.java
@@ -1,0 +1,4 @@
+package com.komsije.booking.e2eTests.pages;
+
+public class GuestProfilePage {
+}
