+++ booking/src/test/java/com/komsije/booking/e2eTests/pages/LoginPage.java
@@ -1,0 +1,4 @@
+package com.komsije.booking.e2eTests.pages;
+
+public class LoginPage {
+}
