+++ booking/src/test/java/com/komsije/booking/e2eTests/tests/CancelReservationTest.java
@@ -1,0 +1,4 @@
+package com.komsije.booking.e2eTests.tests;
+
+public class CancelReservationTest {
+}
