+++ booking/src/test/java/com/komsije/booking/e2eTests/tests/SearchAndFilterTest.java
@@ -1,0 +1,5 @@
+package com.komsije.booking.e2eTests.tests;
+
+public class SearchAndFilterTest {
+
+}
